app/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
pymongo==4.6.1
python-dotenv==1.0.1
bcrypt==4.1.2
motor==3.3.2
pydantic-settings==2.1.0
